% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-edits.R
\name{nettskjema_tsd_apply_edits}
\alias{nettskjema_tsd_apply_edits}
\title{Apply edits to nettskjema data}
\usage{
nettskjema_tsd_apply_edits(
  input_file,
  edits_file,
  output_file = NULL,
  verbose = NULL
)
}
\arguments{
\item{input_file}{string. file path to the raw nettskjema
file, that contains all answers for a form, you want to apply edits to.}

\item{edits_file}{string. file path to the json with
the edits that should be applied}

\item{output_file}{character. Path to output file.}

\item{verbose}{logical. Should the function output status messages. Default is TRUE.}
}
\value{
data.frame with edited nettskjema data
}
\description{
Nettskjema data can at times contain errors
that require manual editing. Editing raw
files directly is not recommended practice,
as logging these changes often is not maintained
over time. Applying edits to nettskjema data
through a more transparent manner means keeping
raw data in its original state, while also
keeping track of the changes needed. This function
needs a json formatted file to apply edits to
either raw csv or json data from nettskjema.
TODO: make vignette with instructions on edit.json
TODO: make functions to initiate/add edits
}
