% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-edits.R
\name{nettskjema_tsd_apply_edits_all}
\alias{nettskjema_tsd_apply_edits_all}
\title{Apply edits to all nettskjema data}
\usage{
nettskjema_tsd_apply_edits_all(
  input_dir = NULL,
  pattern = NULL,
  suffix = "-ed",
  project = NULL,
  output_dir = NULL
)
}
\arguments{
\item{input_dir}{directory path where the input files are}

\item{pattern}{regex pattern to match files with}

\item{project}{current TSD project. Will automatically auto-detect project.}

\item{output_dir}{character. Path to output directory.}
}
\value{
data frame with edits applied
}
\description{
Apply edits to all nettskjema data
}
