% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-cleaners.R
\name{nettskjema_tsd_csv2data}
\alias{nettskjema_tsd_csv2data}
\title{Combine csv's from Nettskjema (on TSD)}
\usage{
nettskjema_tsd_csv2data(input_dir, output_file = NULL)
}
\arguments{
\item{input_dir}{character. Path to input directory.}

\item{output_file}{character. Path to output file.}
}
\value{
data.frame with all data
}
\description{
The submissions received on TSD arrive as both
csv and jsons. Combining all these files
produce the entire data from a form.
This function combines all the csv within TSD
to create a unified data.frame. If \code{output_file}
is provided, will also write the the data for a file.
}
\examples{
\dontrun{
nettskjema_tsd_csv2data("/durable/nettskjema-submissions/cleaned/80902")

nettskjema_tsd_csv2data("/durable/nettskjema-submissions/cleaned/80902",
    output_file = "/durable/nettskjema-submissions/cleaned/80902/data.csv")
}
}
