% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-decrypt.R
\name{nettskjema_tsd_decrypt_all_forms}
\alias{nettskjema_tsd_decrypt_all_forms}
\title{Decrypt all Nettskjema forms}
\usage{
nettskjema_tsd_decrypt_all_forms(
  path = NULL,
  project = NULL,
  output_dir = NULL,
  rerun_all = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{path}{path to the folder containing all the forms}

\item{project}{current TSD project. Will automatically auto-detect project.}

\item{output_dir}{character. Path to output directory.}

\item{rerun_all}{logical. Should the process be run on all files, even those previously run. Default is FALSE.}

\item{verbose}{logical. Should the function output status messages. Default is TRUE.}

\item{...}{other arguments to \code{\link[base]{mapply}}}
}
\value{
success list
}
\description{
Given that all forms are placed in the same basic folder
structures (as they are delivered to TSD), this function
will detect all forms in the path provided and decrypt all
the forms available.
}
