% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-decrypt.R
\name{nettskjema_tsd_decrypt_form}
\alias{nettskjema_tsd_decrypt_form}
\title{Decrypt all files associated with a form}
\usage{
nettskjema_tsd_decrypt_form(path, rerun_all = FALSE, output_dir = NULL, ...)
}
\arguments{
\item{path}{path to the folder with encrypted data}

\item{rerun_all}{logical. Should the process be run on all files, even those previously run. Default is FALSE.}

\item{output_dir}{character. Path to output directory.}

\item{...}{other arguments to \link{nettskjema_tsd_decrypt_file}}
}
\value{
nothing. Writes decrypted files to a folder.
}
\description{
Decrypt all files associated with a form
}
