% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-keyfile.R
\name{nettskjema_tsd_import_key}
\alias{nettskjema_tsd_import_key}
\title{Import GPG key}
\usage{
nettskjema_tsd_import_key(path = NULL, verbose = NULL)
}
\arguments{
\item{path}{full path to the key-file.}

\item{verbose}{logical. Should the function output status messages. Default is TRUE.}
}
\value{
Nothing. Imports the key to the session
}
\description{
As a security measure, all submissions to Nettskjema data in to
TSD are encrypted. These encrypted files need to de decrypted
before they can be sorted and prepared for use in R (or  other
programs). To decrypt the data, R needs to have access to the
secret key that will verify that the decryption is ok to carry
out.
This function will import the necessary key for the decryption
to take place.
If the key-file for the decryption is not automatically detected,
contact IT to learn where your key has been stored. Any R-script
you will run can then specify the location of the key-file either
directly in the call to this function, or as an R-environment variable
'nettskjema.tsd.keyfile=/full/path/to/keyfile'.
}
\examples{
\dontrun{
nettskjema_tsd_import_key()
nettskjema_tsd_import_key(verbose = FALSE)
nettskjema_tsd_import_key(path = "/my/keyfile/path/secret.asc")
}
}
