% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-keyfile.R
\name{nettskjema_tsd_keyfile}
\alias{nettskjema_tsd_keyfile}
\title{Nettskjema TSD keyfile location}
\usage{
nettskjema_tsd_keyfile(
  path = Sys.getenv("nettskjema.tsd.keyfile"),
  project = NULL,
  verbose = NULL
)
}
\arguments{
\item{path}{full path to the key-file}

\item{project}{current TSD project. Will automatically auto-detect project.}

\item{verbose}{logical. Should the function output status messages. Default is TRUE.}
}
\value{
character with the full verified path
}
\description{
As a security measure, all submissions to Nettskjema data in to
TSD are encrypted. These encrypted files need to de decrypted
before they can be sorted and prepared for use in R (or  other
programs). To decrypt the data, R needs to have access to the
secret key that will verify that the decryption is ok to carry
out.
This function can verify the existence of a key-file given a
specified location, or attempt to auto-detect a key-file.
}
\examples{
\dontrun{
nettskjema_tsd_keyfile()
nettskjema_tsd_keyfile(path = "/my/keyfile/path/secret.asc")
}
}
