% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-nettskjema.R
\name{nettskjema_tsd_organise_all_forms}
\alias{nettskjema_tsd_organise_all_forms}
\title{Decrypt all forms associated with a project}
\usage{
nettskjema_tsd_organise_all_forms(rerun_all = FALSE, ...)
}
\arguments{
\item{rerun_all}{logical. Should the process be run on all files, even those previously run. Default is FALSE.}

\item{...}{arguments to pass to \link{nettskjema_tsd_organise_form}}
}
\value{
nothing. Writes decrypted files to a folder.
}
\description{
Decrypt all forms associated with a project
}
