% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-decrypt.R
\name{nettskjema_tsd_submission_path}
\alias{nettskjema_tsd_submission_path}
\title{Find the path for encrypted Nettskjema data}
\usage{
nettskjema_tsd_submission_path(project = NULL)
}
\arguments{
\item{project}{current TSD project. Will automatically auto-detect project.}
}
\value{
single character path
}
\description{
Find the path for encrypted Nettskjema data
}
