% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsd-utils.R
\name{read_nettskjema_tsd}
\alias{read_nettskjema_tsd}
\title{Read nettskjema data}
\usage{
read_nettskjema_tsd(path)
}
\arguments{
\item{path}{string path to a nettskjema file}
}
\value{
a dataframe with nettskjema data
}
\description{
Utility function to read in nettskjema data.
forced all columns to character to not loose leading zero.
}
\examples{
\dontrun{
read_nettskjema_tsd("/tsd/pxx/nettskjema/form.csv")
}
}
